requirements.txt
python-telegram-bot==20.7
Pillow==10.1.0
qrcode==7.4.2
python-dotenv==1.0.0
aiofiles==23.2.1
gunicorn==21.2.0
flask==3.0.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9  # PostgreSQL uchun driver